chore: move access patterns
