+++ packages/@starwars/db/src/lib/access-patterns/__tests__/get-all-entities-for-org.spec.ts
@@ -1,4 +1,4 @@
-import { TableModelsMap } from '../../../types';
+import { TableModelsMap } from '../../types';
 import { getAllEntitiesForOrg } from '../get-all-entities-for-org';
 
 describe('getAllEntitiesForOrg', () => {
