+++ packages/@starwars/db/src/lib/access-patterns/__tests__/get-org.spec.ts
@@ -1,6 +1,6 @@
-import { defaultDbFixture } from '../../../../__fixtures__/default-db.fixture';
-import { createMockModelMap, createMockTable } from '../../../../__test-utils__';
-import { TableModelsMap } from '../../../types';
+import { defaultDbFixture } from '../../../__fixtures__/default-db.fixture';
+import { createMockModelMap, createMockTable } from '../../../__test-utils__';
+import { TableModelsMap } from '../../types';
 import { getOrg } from '../get-org';
 
 describe('getOrg', () => {
