--- packages/@starwars/db/src/lib/access-patterns/admin/index.ts
@@ -1,4 +1,0 @@
-export * from './create-new-user';
-export * from './get-all-entities-for-org';
-export * from './get-org';
-export * from './get-user';
