+++ packages/@starwars/db/src/lib/access-patterns/create-account.ts
@@ -1,0 +1,25 @@
+export const createAccount = {
+  executor: async () => {
+
+  },
+  params: [
+    {
+      name: 'createAccountInput',
+      params: [
+        {
+          name: 'email',
+          type: 'string',
+        },
+        {
+          name: 'accountName',
+          type: 'string',
+        },
+        {
+          name: 'password',
+          type: 'string',
+        },
+      ],
+      type: 'object',
+    },
+  ],
+}
