+++ packages/@starwars/db/src/lib/access-patterns/index.ts
@@ -1,10 +1,4 @@
-import * as admin from './admin';
-import { AccessPatternConfig } from '@ag-oss/one-table';
-
-export * from './admin';
-
-export const gameShopsAccessPatterns = Object.entries(admin).map(([name, config]) => ({
-  executor: config.executor,
-  name: name,
-  params: config.params,
-})) as AccessPatternConfig[];
+export * from './create-new-user';
+export * from './get-all-entities-for-org';
+export * from './get-org';
+export * from './get-user';
